+++ cvd/videobuffer.h
@@ -17,7 +17,7 @@ template <class T> class VideoBuffer
 		virtual VideoFrame<T>* get_frame()=0;        	// blocks until frame ready
 		virtual void put_frame(VideoFrame<T>* f)=0;  	// user is finished using f
 		virtual bool frame_pending()=0;             	// checks to see if frame ready
-		virtual void seek_to(unsigned long long int t)  // in milliseconds
+		virtual void seek_to(double t)					// in seconds
 		{}
 };
